requirements.txt
# Core FastAPI and server
fastapi==0.116.1
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database and Caching
sqlalchemy==2.0.23
alembic==1.12.1
sqlite-utils==3.35.2
redis==5.0.1
aioredis==2.0.1

# Pydantic and settings
pydantic[email]==2.11.7
pydantic-settings==2.10.1

# Local packages
-e file:./trading-engine

# Authentication and Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyotp==2.9.0
webauthn==2.5.1

# Secret Management
hvac>=1.1.0
boto3>=1.26.0
botocore>=1.29.0

# Payment Processing
stripe==7.8.0

# API clients and data
httpx==0.25.2
aiohttp==3.10.11
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0

# Background Tasks
celery==5.3.4

# Data Processing & Analytics
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4
xgboost==2.0.1

# Financial & Trading Libraries
yfinance==0.2.28
alpha-vantage==2.3.1
alpaca-trade-api==3.1.1
ta-lib==0.4.28
quantlib==1.32

# AI/ML Advanced Components
tensorflow==2.15.0
torch==2.1.0
transformers==4.35.0
plotly==5.17.0

# Quantum Computing (Optional - for quantum ML models)
qiskit==0.45.0
qiskit-algorithms==0.2.0
qiskit-machine-learning==0.7.0

# News and Sentiment Analysis
newsapi-python==0.2.6
textblob==0.17.1
vaderSentiment==3.3.2

# Economic Data
pandas-datareader==0.10.0
fredapi==0.5.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==24.8.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Cloud Run specific
gunicorn==21.2.0

# Email validation
email-validator==2.2.0